Move cuda files to separate folder
